+++ src/message/index.ts
@@ -124,12 +124,12 @@ export type { ControlMessage } from "./control";
 
 // AUTHORIZATION_TOKEN Token 構造
 export {
-  type AuthToken,
-  type AuthTokenDelete,
-  type AuthTokenRegister,
-  type AuthTokenUseAlias,
-  type AuthTokenUseValue,
-  AuthTokenAliasType,
-  decodeAuthToken,
-  encodeAuthToken,
-} from "./authToken";
+  type AuthorizationToken,
+  type AuthorizationTokenDelete,
+  type AuthorizationTokenRegister,
+  type AuthorizationTokenUseAlias,
+  type AuthorizationTokenUseValue,
+  AuthorizationTokenAliasType,
+  decodeAuthorizationToken,
+  encodeAuthorizationToken,
+} from "./authorizationToken";